Include version info in the RDF.
Define an IRI namespace under an Apollo Protocol domain to represent
this library. Include relations to subjects under that namespace to
indicate the version of the library and the version of the HDQM->RDF
representation used.

Implement this via a general method to include versioning triples in the
data store, as ActivityLib will need to include separate versioning
information. Any future refactoring to represent the data with more
specific classes than Thing may need to arrange to store this version
data separately from the HQDM data.

For now accept unversioned files and assume they conform to the version
1 representation. This may want to be changed in the future.